Added new PdfContent class which wraps the complexities of content stream vs. array Modified PdfCanvas, PdfPage, PdfXObject & PdfPainter to now do the right thing - so we now support drawing on any page! PdfObjects retrieved from GetIndirectKey() have a valid m_pParent - regardless of direct/indirect state Fixed a bunch of compiler warnings Fixed bugs introduced in PdfDocument with PdfObject* vs. PdfObject
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@187 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/PdfStream.cpp
@@ -83,7 +83,7 @@ void PdfStream::Append( const char* pszString )
     this->Append( pszString, strlen( pszString ) );
 }
 
-void PdfStream::Append( const char* pszString, long lLen )
+void PdfStream::Append( const char* pszString, size_t lLen )
 {
     if( !pszString )
     {
